Consider torch libraries as system libraries
Fix for making `delocate` consider the `torch` libraries as system
libraries, excluding them for being vendored into the resulting wheels.
This is an attempt to follow torch's recommendation [1] of relying on
`import torch` for loading their libraries and resolving the symbols at
runtime, so subsequent libraries that depend on them can be loaded
correctly (without needing a separate copy of the torch libraries).

[1] https://pytorch.org/tutorials/advanced/cpp_extension.html#using-your-extension